*** QSD *** -> problem with QSD Act 025 fixed, clan point structure was wrong since operator was wrong...,
*** Clan Points ***
-> packet sent to Char Server wasn't good sometimes to update clan points amount,

+++ World Server/quest/QuestActions.h
@@ -220,8 +220,8 @@ struct STR_REWD_024 {
 };
 
 struct STR_REWD_025 {
-	short nPOINT;
-	//int nPOINT;
+	//short nPOINT;
+	int nPOINT;
 	byte btOP;
 };
